backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
arxiv>=2.1.3
openai>=1.51.0
anthropic>=0.34.0
google-genai>=0.1.2
httpx>=0.27.2
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
scikit-learn>=1.6.0
numpy>=2.2.0
aiofiles>=24.1.0
python-dotenv>=1.0.1
apscheduler>=3.10.4
tenacity>=8.5.0
pymupdf>=1.24.0
